requirements.txt
allure-pytest
playwright
pytest
pytest-asyncio
pylint-fail-under
pytest-xdist
